Restructure: Move Wiznet PIO SPI code to new wiznet module
