+++ ports/raspberrypi/Makefile
@@ -83,13 +83,13 @@ SRC_CYW43 :=
 SRC_LWIP :=
 endif
 
-ifeq ($(CIRCUITPY_WIZNET_PIO_SPI), 1)
-INC_WIZNET_PIO_SPI := \
-	-isystem wiznet_pio_spi \
+ifeq ($(CIRCUITPY_WIZNET), 1)
+INC_WIZNET := \
+	-isystem common-hal/wiznet \
 
-SRC_WIZNET_PIO_SPI := \
-	wiznet_pio_spi/wizchip_pio_spi.c \
-
+SRC_WIZNET := \
+	common-hal/wiznet/wizchip_pio_spi.c \
+	
 WIZNET_PIOASM = $(BUILD)/pioasm/pioasm/pioasm
 .PHONY: pioasmBuild
 pioasmBuild: $(WIZNET_PIOASM)
@@ -98,15 +98,15 @@ $(WIZNET_PIOASM):
 	$(Q)cmake -S pioasm -B $(BUILD)/pioasm
 	$(MAKE) -C $(BUILD)/pioasm pioasmBuild
 
-$(BUILD)/wizchip_pio_spi.pio.h: wiznet_pio_spi/wizchip_pio_spi.pio $(WIZNET_PIOASM)
+$(BUILD)/wizchip_pio_spi.pio.h: common-hal/wiznet/wizchip_pio_spi.pio $(WIZNET_PIOASM)
 	$(Q)$(WIZNET_PIOASM) -o c-sdk $< $@
-$(BUILD)/wiznet_pio_spi/wizchip_pio_spi.o: $(BUILD)/wizchip_pio_spi.pio.h
+$(BUILD)/common-hal/wiznet/wizchip_pio_spi.o: $(BUILD)/wizchip_pio_spi.pio.h
 
 $(BUILD)/genhdr/qstr.i.last: $(BUILD)/wizchip_pio_spi.pio.h
 
 else
-INC_WIZNET_PIO_SPI :=
-SRC_WIZNET_PIO_SPI :=
+INC_WIZNET :=
+SRC_WIZNET :=
 
 endif
 
@@ -182,7 +182,6 @@ INC += \
 	-isystem sdk/src/rp2_common/pico_time_adapter/include/ \
 	-isystem sdk/src/rp2_common/pico_unique_id/include/ \
 	$(INC_CYW43) \
-	$(INC_WIZNET_PIO_SPI) \
 	-Isdk_config \
 	-I../../lib/tinyusb/src \
 	-I../../supervisor/shared/usb \
@@ -560,6 +559,10 @@ SRC_C += \
 	bindings/rp2pio/__init__.c \
 	common-hal/rp2pio/StateMachine.c \
 	common-hal/rp2pio/__init__.c \
+	bindings/wiznet/PIO_SPI.c \
+	bindings/wiznet/__init__.c \
+	common-hal/wiznet/PIO_SPI.c \
+	common-hal/wiznet/__init__.c \
 	audio_dma.c \
 	background.c \
 	peripherals/pins.c \
@@ -569,7 +572,7 @@ SRC_C += \
 	mphalport.c \
 	$(SRC_CYW43) \
 	$(SRC_LWIP) \
-	$(SRC_WIZNET_PIO_SPI) \
+	$(SRC_WIZNET) \
 
 
 ifeq ($(CIRCUITPY_USB_HOST), 1)
