+++ ports/raspberrypi/bindings/wiznet/__init__.h
@@ -1,0 +1,7 @@
+// This file is part of the CircuitPython project: https://circuitpython.org
+//
+// SPDX-FileCopyrightText: Copyright (c) 2016 Scott Shawcroft
+//
+// SPDX-License-Identifier: MIT
+
+#pragma once
