+++ ports/raspberrypi/boards/wiznet_w55rp20_evb_pico/mpconfigboard.mk
@@ -6,9 +6,9 @@ USB_MANUFACTURER = "WIZnet"
 CHIP_VARIANT = RP2040
 CHIP_FAMILY = rp2
 
-EXTERNAL_FLASH_DEVICES = "W25Q16JVxQ"
+EXTERNAL_FLASH_DEVICES = "PY25Q16HB"
 
-CIRCUITPY_WIZNET_PIO_SPI = 1
+CIRCUITPY_WIZNET = 1
 CIRCUITPY__EVE = 1
 CIRCUITPY_SSL = 1
 CIRCUITPY_USB_HOST = 0
