+++ ports/raspberrypi/common-hal/wiznet/__init__.c
@@ -1,0 +1,7 @@
+// This file is part of the CircuitPython project: https://circuitpython.org
+//
+// SPDX-FileCopyrightText: Copyright (c) 2024 Adafruit Industries LLC
+//
+// SPDX-License-Identifier: MIT
+
+// No wiznet module functions.
