+++ py/circuitpy_mpconfig.mk
@@ -206,9 +206,6 @@ CFLAGS += -DCIRCUITPY_BUSIO_SPI=$(CIRCUITPY_BUSIO_SPI)
 CIRCUITPY_BUSIO_UART ?= $(CIRCUITPY_BUSIO)
 CFLAGS += -DCIRCUITPY_BUSIO_UART=$(CIRCUITPY_BUSIO_UART)
 
-CIRCUITPY_WIZNET_PIO_SPI ?= 0
-CFLAGS += -DCIRCUITPY_WIZNET_PIO_SPI=$(CIRCUITPY_WIZNET_PIO_SPI)
-
 CIRCUITPY_CAMERA ?= 0
 CFLAGS += -DCIRCUITPY_CAMERA=$(CIRCUITPY_CAMERA)
 
@@ -706,6 +703,9 @@ CFLAGS += -DCIRCUITPY_WATCHDOG=$(CIRCUITPY_WATCHDOG)
 
 CIRCUITPY_WIFI ?= 0
 CFLAGS += -DCIRCUITPY_WIFI=$(CIRCUITPY_WIFI)
+
+CIRCUITPY_WIZNET ?= 0
+CFLAGS += -DCIRCUITPY_WIZNET=$(CIRCUITPY_WIZNET)
 
 CIRCUITPY_WEB_WORKFLOW ?= $(CIRCUITPY_WIFI)
 CFLAGS += -DCIRCUITPY_WEB_WORKFLOW=$(CIRCUITPY_WEB_WORKFLOW)
